requirements.txt
yfinance==0.2.32
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.1
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
joblib==1.3.2
pyyaml==6.0.1
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
jupyter==1.0.0
ipython==8.19.0

